# dcm_common/orchestra/models/report.py
"""Definition of the `Report`-model."""

from typing import Optional
from dataclasses import dataclass, field
from enum import Enum

from dcm_common import Logger
from dcm_common.models import DataModel, JSONable
from .token import Token


class Status(Enum):
    """
    Enum-class for different types of job-status.
    """

    QUEUED = "queued"
    RUNNING = "running"
    ABORTED = "aborted"
    COMPLETED = "completed"


@dataclass
class Progress(DataModel):
    """
    Progress `DataModel`

    Keyword arguments:
    status -- `Job`'s `Status` in processing pipeline
    verbose -- verbose description of `Job`-progress
    numeric -- numeric representation of `Progress` in percent
    """

    status: Status = Status.QUEUED
    verbose: str = field(default_factory=lambda: "")
    numeric: int = 0

    @DataModel.serialization_handler("status")
    @classmethod
    def status_serialization(cls, value):
        """Performs `status`-serialization."""
        return value.value

    @DataModel.deserialization_handler("status")
    @classmethod
    def status_deserialization(cls, value):
        """Performs `status`-deserialization."""
        return Status(value)

    def run(self) -> None:
        """Set `status`-property to RUNNING."""
        self.status = Status.RUNNING

    def queue(self) -> None:
        """Set `status`-property to QUEUED."""
        self.status = Status.QUEUED

    def abort(self) -> None:
        """Set `status`-property to ABORTED."""
        self.status = Status.ABORTED

    def complete(self) -> None:
        """Set `status`-property to COMPLETED."""
        self.status = Status.COMPLETED

    def create_verbose_update_callback(
        self, prefix: Optional[str] = None
    ) -> None:
        """Returns callback to set `verbose`-property."""
        if prefix is None:
            _prefix = ""
        else:
            _prefix = f"{prefix}: "
        return lambda value: setattr(self, "verbose", _prefix + value)


@dataclass(kw_only=True)
class Report(DataModel):
    """
    Report `DataModel`

    Define child models based on `Report` by, for example, writing
     >>> @dataclass
     ... class ReportB(Report):
     ...    data: str

    Keyword arguments:
    host -- service url where this report has been generated
    token -- `Job`-token that is associated with this report
    args -- requestBody for this `Job`
    progress -- `Job` progress record
    log -- `Logger` containing `Job` related information
    """

    host: Optional[str] = field(default_factory=lambda: "unknown")
    token: Optional[Token] = None
    args: JSONable = None
    progress: Progress = field(default_factory=Progress)
    log: Logger = field(default_factory=Logger)