# dcm_common/services/demo/openapi.yaml
openapi: 3.0.3
info:
  title: DCM Demo Service - OpenAPI-Document
  description: |-
    This is the OpenAPI 3.0 specification for a generic DCM service used
    for tests.
  version: 0.2.0
servers:
  - url: 'http://localhost:8080/'
...
paths:
...
  /status:
    get:
      summary: container status
      description: Returns Status (readiness) of container.
      operationId: get_status
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        '400':
          $ref: '#/components/responses/400BadRequest'
  /identify:
    get:
      summary: container self-description
      description: Returns a detailed description of the container properties.
      operationId: identify
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SelfDescription'
        '400':
          $ref: '#/components/responses/400BadRequest'
...
components:
  schemas:
    Status:
      title: ContainerStatus
      description:
        A ContainerStatus gives a summary of the container's status like
        availability for job submissions
      type: object
      properties:
        ready:
          type: boolean
          description: availability of container for a request
          example: true
      required:
        - ready
...
    PluginArgSignature:
      title: PluginArgSignature
      type: object
      description: plugin argument
      properties:
        type:
          type: string
          description: JSON type of the argument
          enum:
            - "string"
            - "integer"
            - "number"
            - "boolean"
            - "array"
            - "object"
          example: string
        required:
          type: boolean
          description:
            if true, this argument is required for a valid request
          example: false
        description:
          type: string
          description: description of argument's purpose
          example: this argument sets ...
        example:
          description: example for argument value
          example: 0
        default:
          description: |-
            default value

            only applicable in case of "required: false" and only to
            primitive types or array
          example: 0
        itemType:
          oneOf:
            - title: PrimitiveType
              type: string
              enum:
                - "string"
                - "integer"
                - "number"
                - "boolean"
              description: |-
                required in case of "type: array"; type of array elements

                Note that all arguments in an array are required to have the
                same type and cannot have default values assigned individually.
                A default can be given for the entire array in the parent
                PluginArgSignature. Only primitive types are allowed.
            - $ref: '#/components/schemas/PluginArgSignature'
        properties:
          type: object
          description: |-
            required in case of "type: object"; mapping of child-names
            and child-PluginArgSignature
          additionalProperties:  # valid but not rendered in SwaggerUI
            $ref: '#/components/schemas/PluginArgSignature'
        additional_properties:
          type: boolean
          description: |-
            only applicable in case of "type: object"; if false, only
            explicitly listed properties are accepted
          example: false
      example:
        type: integer
        required: false
        default: 0
      required:
        - type
        - required
    Plugin:
      title: Plugin
      description: DCM-service plugin
      type: object
      properties:
        name:
          type: string
          description: plugin identifier
          example: plugin-0
        description:
          type: string
          description: description of plugin features
          example: performs ...
        context:
          type: string
          description: identifier for the plugin context
          example: context-0
        signature:
          $ref: '#/components/schemas/PluginArgSignature'
        dependencies:
          type: object
          description:
            mapping of names and installed versions of plugin dependencies
          additionalProperties:
            type: string
            description: version string
            example: 1.0.0
          example:
            python-package: 1.0.0
        info:
          type: object
          description: additional information
          additionalProperties: true
      required:
        - name
        - description
        - signature
...
    SelfDescription:
      type: object
      description: information regarding the service's properties
      properties:
        description:
          type: string
          description:
            a verbose self-description of the service functionality
          example:
            This service provides endpoints for ...
        version:
          type: object
          description: collection of version strings
          properties:
            api:
              type: string
              description: version string of the implemented API
              example: "0.2.0"
            app:
              type: string
              description: version string of the app
              example: "0.1.0"
            python:
              type: string
              description: version of the python implementation
              example: 3.10.12 (main, Mar 22 2024, 16:50:05) [GCC 11.4.0]
            software:
              type: object
              description: deployed software versions
            lib:
              type: object
              description: deployed library versions
              additionalProperties:
                type: string
                description: version string
                example: "0.1.0"
              example:
                dependency-a: "1.0.0"
                dependency-b: "0.2.0.post1"
          required:
            - api
            - app
            - python
            - software
            - lib
        configuration:
          type: object
          description: service configuration details
          properties:
            settings:
              type: object
              properties:
                allow_cors:
                  type: boolean
                  description: whether CORS requests are allowed
                  example: true
                fs_mount_point:
                  type: string
                  description: mount point of shared file system
                  example: /file_storage
                orchestra:
                  type: object
                  description: orchestra-related settings
                  properties:
                    atStartup:
                      type: boolean
                      description:
                        whether orchestration-processing loop is run at startup
                      example: true
                    poolSize:
                      type: integer
                      description: number of worker-threads
                      example: 1
                    daemonInterval:
                      type: number
                      description: interval of orchestra-restarts in seconds
                      example: 1
                    controller:
                      type: object
                      description: controller settings
                      properties:
                        type:
                          type: string
                          description: controller type
                          example: sqlite
                        args:
                          type: object
                          description: controller arguments
                          additionalProperties: true
                          example:
                            lock_ttl: 30
                      required:
                        - type
                        - args
                    worker:
                      type: object
                      description: worker settings
                      properties:
                        args:
                          type: object
                          description: controller arguments
                          additionalProperties: true
                          example:
                            lock_refresh_interval: 10
                        interval:
                          type: number
                          description: interval of the worker loop in seconds
                          example: 1
                      required:
                        - args
                        - interval
                    abortTimeout:
                      type: number
                      description: duration for a timeout in an abort-request
                  required:
                    - atStartup
                    - poolSize
                    - daemonInterval
                    - controller
                    - worker
                    - abortTimeout
                database:
                  type: object
                  description:
                    database-related settings
                  properties:
                    adapter:
                      type: string
                      description: database adapter name
                      example: postgres
                    schemaVersion:
                      type: string
                      description: dcm-database schema version string
                      example: "1.0.0"
                    settings:
                      type: object
                      description: database-related settings
                      properties:
                        poolSize:
                          type: number
                          description: size of database connection-pool
                          example: 10
                        poolOverflow:
                          type: boolean
                          description:
                            whether to allow the pool to overflow during
                            high load (dynamically allocates more connections
                            when needed)
                          example: true
                        connectionTimeout:
                          type: number
                          nullable: true
                          description:
                            duration after which a database connection-
                            attempt times out; a value of null corresponds
                            to no timeout
                          example: 10.0
                        file:
                          type: string
                          nullable: true
                          description:
                            file path for a SQLite database (relative to
                            working directory); null corresponds to an
                            in-memory database
                          example: sqlite.db
                        host:
                          type: string
                          description: PostgreSQL host name
                          example: database
                        name:
                          type: string
                          description: PostgreSQL database name
                          example: dcm-database
                        user:
                          type: string
                          description: PostgreSQL user name
                          example: dcm
                        port:
                          type: string
                          description: PostgreSQL port
                          example: 5432
                        password:
                          type: boolean
                          description: whether PostgreSQL password is set
                          example: true
                        passfile:
                          type: string
                          description: PostgreSQL passfile path
                          example: pass
              required:
                - allow_cors
                - orchestra
                - database
            plugins:
              type: object
              description: mapping of available plugins
              additionalProperties:
                $ref: "#/components/schemas/Plugin"
              example:
                plugin-0:
                  name: plugin-0
                  description: performs ...
                  context: context-0
                  signature:
                    type: object
                    required: true
                    properties:
                      param-0:
                        type: integer
                        required: false
                        default: 0
                    additional_properties: false
            services:
              type: object
              description:
                mapping of hostname and url for secondary DCM-services
              additionalProperties:
                title: Service
                type: string
                format: url
              example:
                service: <service-url>
          required:
            - settings
            - plugins
            - services
      required:
        - description
        - version
        - configuration
...
  responses:
...
    400BadRequest:
      description: malformed request body
      content:
        text/plain:
          schema:
            type: string
            example: "Missing required argument <X>"
...